+++ src/main/java/models/ContaCorrente.java
@@ -7,7 +7,7 @@ public class ContaCorrente extends ContaBancaria
     private double valorCestaServicos;
     private double limitePixNoturno;
 
-    public ContaCorrente(int id, int codigoBanco, int numAgencia, int numConta, double saldo, Date dataAbertura, int idTitular, double valorCestaServicos, double limitePixNoturno)
+    public ContaCorrente(long id, long codigoBanco, int numAgencia, long numConta, double saldo, Date dataAbertura, long idTitular, double valorCestaServicos, double limitePixNoturno)
     {
         super(id, codigoBanco, numAgencia, numConta, saldo, dataAbertura, idTitular);
         setValorCestaServicos(valorCestaServicos);
